Converted client tab complete packet
